Deduplicate metrics loading in tests.
